requirements.txt
fastapi
uvicorn[standard]
tensorflow==2.11.0
numpy==1.24.4
pillow
python-multipart
requests
